pyproject.toml
[project]
name = "skill-cli"
version = "0.1.0"
description = "从 Git 仓库按组安装 Claude Code Skills"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
skill-cli = "skill_cli.cli:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
